Feat: 검색/검색 결과 화면 -  Base 코드 적용, View-Controller 분리
